# Minor: parallelize cli part validation test