- This is a small follow on to https://github.com/clflushopt/tpchgen-rs/pull/156

I noticed it took a while for the cli test to run, so let's use multiple threads

Before (15.6 s)

```
(venv) andrewlamb@Andrews-MacBook-Pro-3:~/Software/tpchgen-rs$ cargo nextest  run --test cli_integration
   Compiling tpchgen-cli v1.1.1 (/Users/andrewlamb/Software/tpchgen-rs/tpchgen-cli)
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.47s
────────────
 Nextest run ID 18ce3d74-1627-4f9c-b38d-4bf935ff9959 with nextest profile: default
    Starting 3 tests across 1 binary
        PASS [   4.325s] tpchgen-cli::cli_integration test_tpchgen_cli_tbl_scale_factor_0_001
        PASS [   8.126s] tpchgen-cli::cli_integration test_write_parquet_orders
        PASS [  15.685s] tpchgen-cli::cli_integration test_tpchgen_cli_parts
```

After (4.2 s)
```
(venv) andrewlamb@Andrews-MacBook-Pro-3:~/Software/tpchgen-rs$ cargo nextest  run --test cli_integration
   Compiling tpchgen-cli v1.1.1 (/Users/andrewlamb/Software/tpchgen-rs/tpchgen-cli)
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.40s
────────────
 Nextest run ID 7d5dedd9-35cb-4dd8-ae64-11a3807411de with nextest profile: default
    Starting 3 tests across 1 binary
        PASS [   4.297s] tpchgen-cli::cli_integration test_tpchgen_cli_parts
        PASS [   4.325s] tpchgen-cli::cli_integration test_tpchgen_cli_tbl_scale_factor_0_001
        PASS [   8.063s] tpchgen-cli::cli_integration test_write_parquet_orders
────────────
```