@alamb 2025-07-21
It would be nice to run also parallelize the parquet write test, but I tried and did not get anything