@alamb 2025-07-23
Of course once we figure out how to parallelize the individual part generation this might not be necessary, but for now let's make CI faster